test(tau-coding-agent): split runtime_and_startup tests (#1377)
